Clean Up
InteropContract remove unnecessary methods.
Moved Nexus to a specific folder to use Virtual Nexus.
Moved Chain to Specific Folder.
Added Comments to Runtime.
