+++ Phantasma.Business/src/Blockchain/Chains/ChainTasks.cs
@@ -1,0 +1,6 @@
+﻿namespace Phantasma.Business.Blockchain;
+
+public  class ChainTasks
+{
+    
+}
