+++ Phantasma.Business/src/Blockchain/VM/Runtime/RuntimeTriggers.cs
@@ -21,206 +21,233 @@ public partial class RuntimeVM: GasMachine, IRuntime
 {
      private HashSet<string> _triggerGuards = new HashSet<string>();
 
-        internal void ValidateTriggerGuard(string triggerName)
-        {
-            ExpectNameLength(triggerName, nameof(triggerName));
+     /// <summary>
+     /// Validate Trigger Guard
+     /// </summary>
+     /// <param name="triggerName"></param>
+     /// <exception cref="ChainException"></exception>
+    internal void ValidateTriggerGuard(string triggerName)
+    {
+        ExpectNameLength(triggerName, nameof(triggerName));
 
-            if (ProtocolVersion <= DomainSettings.Phantasma30Protocol)
-            {
-                if (_triggerGuards.Contains(triggerName))
-                {
-                    throw new ChainException("trigger loop detected: " + triggerName);
-                }
-            }
-            else if (_triggerGuards.Count >= DomainSettings.MaxTriggerLoop)
+        if (ProtocolVersion <= DomainSettings.Phantasma30Protocol)
+        {
+            if (_triggerGuards.Contains(triggerName))
             {
                 throw new ChainException("trigger loop detected: " + triggerName);
             }
-
-
-            _triggerGuards.Add(triggerName);
+        }
+        else if (_triggerGuards.Count >= DomainSettings.MaxTriggerLoop)
+        {
+            throw new ChainException("trigger loop detected: " + triggerName);
         }
 
-        #region TRIGGERS
 
-        public TriggerResult InvokeTriggerOnContract(bool allowThrow, Address address, ContractTrigger trigger,
-            params object[] args)
+        _triggerGuards.Add(triggerName);
+    }
+
+    /// <summary>
+    /// Invoke Trigger on Contract
+    /// </summary>
+    /// <param name="allowThrow"></param>
+    /// <param name="address"></param>
+    /// <param name="trigger"></param>
+    /// <param name="args"></param>
+    /// <returns></returns>
+    public TriggerResult InvokeTriggerOnContract(bool allowThrow, Address address, ContractTrigger trigger,
+        params object[] args)
+    {
+        ExpectAddressSize(address, nameof(address));
+        ExpectEnumIsDefined(trigger, nameof(trigger));
+        ExpectArgsLength(args, nameof(args));
+
+        if (address.IsNull)
         {
-            ExpectAddressSize(address, nameof(address));
-            ExpectEnumIsDefined(trigger, nameof(trigger));
-            ExpectArgsLength(args, nameof(args));
+            return TriggerResult.Failure;
+        }
 
-            if (address.IsNull)
+        if (!IsTrigger)
+        {
+            _triggerGuards.Clear();
+        }
+
+        var triggerName = trigger.ToString();
+
+        if (address.IsUser)
+        {
+            //var accountScript = Nexus.LookUpAddressScript(RootStorage, address);
+            var accountABI = OptimizedAddressABILookup(address);
+            var accountScript = accountABI != null ? OptimizedAddressScriptLookup(address) : null;
+
+            //Expect(accountScript.SequenceEqual(accountScript2), "different account scripts");
+
+            ValidateTriggerGuard($"{address.Text}.{triggerName}");
+            return InvokeTrigger(allowThrow, accountScript, address.Text, accountABI, triggerName, args);
+        }
+
+        if (address.IsSystem)
+        {
+            var contract = Chain.GetContractByAddress(Storage, address);
+            if (contract != null)
             {
-                return TriggerResult.Failure;
-            }
-
-            if (!IsTrigger)
-            {
-                _triggerGuards.Clear();
-            }
-
-            var triggerName = trigger.ToString();
-
-            if (address.IsUser)
-            {
-                //var accountScript = Nexus.LookUpAddressScript(RootStorage, address);
-                var accountABI = OptimizedAddressABILookup(address);
-                var accountScript = accountABI != null ? OptimizedAddressScriptLookup(address) : null;
-
-                //Expect(accountScript.SequenceEqual(accountScript2), "different account scripts");
-
-                ValidateTriggerGuard($"{address.Text}.{triggerName}");
-                return InvokeTrigger(allowThrow, accountScript, address.Text, accountABI, triggerName, args);
-            }
-
-            if (address.IsSystem)
-            {
-                var contract = Chain.GetContractByAddress(Storage, address);
-                if (contract != null)
+                if (contract.ABI.HasMethod(triggerName))
                 {
-                    if (contract.ABI.HasMethod(triggerName))
+                    var customContract = contract as CustomContract;
+                    if (customContract != null)
                     {
-                        var customContract = contract as CustomContract;
-                        if (customContract != null)
+                        ValidateTriggerGuard($"{contract.Name}.{triggerName}");
+                        return InvokeTrigger(allowThrow, customContract.Script, contract.Name, contract.ABI,
+                            triggerName, args);
+                    }
+
+                    var native = contract as NativeContract;
+                    if (native != null)
+                    {
+                        ValidateTriggerGuard($"{contract.Name}.{triggerName}");
+
+                        try
                         {
-                            ValidateTriggerGuard($"{contract.Name}.{triggerName}");
-                            return InvokeTrigger(allowThrow, customContract.Script, contract.Name, contract.ABI,
-                                triggerName, args);
+                            this.CallNativeContext(native.Kind, triggerName, args);
+                            return TriggerResult.Success;
                         }
-
-                        var native = contract as NativeContract;
-                        if (native != null)
+                        catch
                         {
-                            ValidateTriggerGuard($"{contract.Name}.{triggerName}");
-
-                            try
+                            if (allowThrow)
                             {
-                                this.CallNativeContext(native.Kind, triggerName, args);
-                                return TriggerResult.Success;
+                                // just throw to preserve original stack trace
+                                throw;
                             }
-                            catch
-                            {
-                                if (allowThrow)
-                                {
-                                    // just throw to preserve original stack trace
-                                    throw;
-                                }
 
-                                return TriggerResult.Failure;
-                            }
+                            return TriggerResult.Failure;
                         }
                     }
                 }
-                else
-                {
-                    return TriggerResult.Missing;
-                }
             }
+            else
+            {
+                return TriggerResult.Missing;
+            }
+        }
 
+        return TriggerResult.Missing;
+    }
+
+    private static byte[] _optimizedScriptMapKey = null;
+    private static byte[] _optimizedABIMapKey = null;
+
+    private static void PrepareOptimizedScriptMapKey()
+    {
+        if (_optimizedScriptMapKey == null)
+        {
+            var accountContractName = NativeContractKind.Account.GetContractName();
+            _optimizedScriptMapKey = Encoding.UTF8.GetBytes($".{accountContractName}._scriptMap");
+        }
+    }
+
+    private byte[] OptimizedAddressScriptLookup(Address target)
+    {
+        PrepareOptimizedScriptMapKey();
+
+        var scriptMap = new StorageMap(_optimizedScriptMapKey, RootStorage);
+
+        if (scriptMap.ContainsKey(target))
+            return scriptMap.Get<Address, byte[]>(target);
+        else
+            return new byte[0];
+    }
+
+    /// <summary>
+    /// Invoke Trigger on Token
+    /// </summary>
+    /// <param name="allowThrow"></param>
+    /// <param name="token"></param>
+    /// <param name="trigger"></param>
+    /// <param name="args"></param>
+    /// <returns></returns>
+    public TriggerResult InvokeTriggerOnToken(bool allowThrow, IToken token, TokenTrigger trigger,
+        params object[] args)
+    {
+        ExpectArgsLength(args, nameof(args));
+        ExpectValidToken(token);
+        ExpectEnumIsDefined(trigger, nameof(trigger));
+        ExpectArgsLength(args, nameof(args));
+
+        return InvokeTrigger(allowThrow, token.Script, token.Symbol, token.ABI, trigger.ToString(), args);
+    }
+
+    /// <summary>
+    /// Invoke Trigger
+    /// </summary>
+    /// <param name="allowThrow"></param>
+    /// <param name="script"></param>
+    /// <param name="contextName"></param>
+    /// <param name="abi"></param>
+    /// <param name="triggerName"></param>
+    /// <param name="args"></param>
+    /// <returns></returns>
+    /// <exception cref="Exception"></exception>
+    public TriggerResult InvokeTrigger(bool allowThrow, byte[] script, string contextName, ContractInterface abi,
+        string triggerName, params object[] args)
+    {
+        ExpectNameLength(contextName, nameof(contextName));
+        ExpectNameLength(triggerName, nameof(triggerName));
+        ExpectArgsLength(args, nameof(args));
+
+        if (script == null || script.Length == 0 || abi == null)
+        {
             return TriggerResult.Missing;
         }
 
-        private static byte[] _optimizedScriptMapKey = null;
-        private static byte[] _optimizedABIMapKey = null;
-
-        private static void PrepareOptimizedScriptMapKey()
+        var method = abi.FindMethod(triggerName);
+        if (method == null || method.offset < 0)
         {
-            if (_optimizedScriptMapKey == null)
-            {
-                var accountContractName = NativeContractKind.Account.GetContractName();
-                _optimizedScriptMapKey = Encoding.UTF8.GetBytes($".{accountContractName}._scriptMap");
-            }
+            return TriggerResult.Missing;
         }
 
-        private byte[] OptimizedAddressScriptLookup(Address target)
+        var runtime = new RuntimeVM(-1, script, (uint)method.offset, Chain, Validator, Time, Transaction, changeSet,
+            Oracle, ChainTask.Null, true, contextName, this);
+
+        for (int i = args.Length - 1; i >= 0; i--)
         {
-            PrepareOptimizedScriptMapKey();
-
-            var scriptMap = new StorageMap(_optimizedScriptMapKey, RootStorage);
-
-            if (scriptMap.ContainsKey(target))
-                return scriptMap.Get<Address, byte[]>(target);
-            else
-                return new byte[0];
+            var obj = VMObject.FromObject(args[i]);
+            runtime.Stack.Push(obj);
         }
 
-        public TriggerResult InvokeTriggerOnToken(bool allowThrow, IToken token, TokenTrigger trigger,
-            params object[] args)
+        ExecutionState state;
+        try
         {
-            ExpectArgsLength(args, nameof(args));
-            ExpectValidToken(token);
-            ExpectEnumIsDefined(trigger, nameof(trigger));
-            ExpectArgsLength(args, nameof(args));
+            state = runtime.Execute();
+            // TODO catch VM exceptions?
+        }
+        catch (VMException)
+        {
+            if (allowThrow)
+            {
+                throw;
+            }
 
-            return InvokeTrigger(allowThrow, token.Script, token.Symbol, token.ABI, trigger.ToString(), args);
+            state = ExecutionState.Fault;
         }
 
-        public TriggerResult InvokeTrigger(bool allowThrow, byte[] script, string contextName, ContractInterface abi,
-            string triggerName, params object[] args)
+        if (state == ExecutionState.Halt)
         {
-            ExpectNameLength(contextName, nameof(contextName));
-            ExpectNameLength(triggerName, nameof(triggerName));
-            ExpectArgsLength(args, nameof(args));
-
-            if (script == null || script.Length == 0 || abi == null)
+            // propagate events to the other runtime
+            foreach (var evt in runtime.Events)
             {
-                return TriggerResult.Missing;
+                Notify(evt.Kind, evt.Address, evt.Data, evt.Contract);
             }
 
-            var method = abi.FindMethod(triggerName);
-            if (method == null || method.offset < 0)
-            {
-                return TriggerResult.Missing;
-            }
-
-            var runtime = new RuntimeVM(-1, script, (uint)method.offset, Chain, Validator, Time, Transaction, changeSet,
-                Oracle, ChainTask.Null, true, contextName, this);
-
-            for (int i = args.Length - 1; i >= 0; i--)
-            {
-                var obj = VMObject.FromObject(args[i]);
-                runtime.Stack.Push(obj);
-            }
-
-            ExecutionState state;
-            try
-            {
-                state = runtime.Execute();
-                // TODO catch VM exceptions?
-            }
-            catch (VMException)
-            {
-                if (allowThrow)
-                {
-                    throw;
-                }
-
-                state = ExecutionState.Fault;
-            }
-
-            if (state == ExecutionState.Halt)
-            {
-                // propagate events to the other runtime
-                foreach (var evt in runtime.Events)
-                {
-                    Notify(evt.Kind, evt.Address, evt.Data, evt.Contract);
-                }
-
-                return TriggerResult.Success;
-            }
-            else
-            {
-                if (allowThrow)
-                {
-                    var vmException = runtime.Stack.Pop().AsString();
-                    throw new Exception($"{triggerName} trigger failed: {vmException}");
-                }
-
-                return TriggerResult.Failure;
-            }
+            return TriggerResult.Success;
         }
+        else
+        {
+            if (allowThrow)
+            {
+                var vmException = runtime.Stack.Pop().AsString();
+                throw new Exception($"{triggerName} trigger failed: {vmException}");
+            }
 
-        #endregion
-
+            return TriggerResult.Failure;
+        }
+    }
 }
