+++ Phantasma.Node/src/Node.cs
@@ -135,7 +135,7 @@ namespace Phantasma.Node
             
             Log.Information($"Server is up & running ()");
 
-            SetupOracleApis();
+            //SetupOracleApis();
 
             SetupNexusApi();
